requirements.txt
 
langchain
langgraph
google-generativeai
pandas
numpy
seaborn
matplotlib
openpyxl
streamlit
python-dotenv
scikit-learn 
plotly        
tqdm
